Move JSON handlers to utility file
