+++ flake8_test
@@ -1,1 +1,1 @@
-flake8 util/save.py
+flake8 util/*.py
